Add sensor data gathering activity

+++ stepDetector/src/main/java/com/keepfit/stepdetection/accelerometer/filter/AccelerometerFilter.java
@@ -1,5 +1,5 @@
 package com.keepfit.stepdetection.accelerometer.filter;
 
 public interface AccelerometerFilter {
-    float[] filter(float x, float y, float z);
+    double[] filter(double x, double y, double z);
 }
